add a new controller classes and fix the some problems

+++ src/main/java/com/diplomado/springboot/repositories/AccomplishmentsRepository.java
@@ -2,6 +2,7 @@ package com.diplomado.springboot.repositories;
 
 import com.diplomado.springboot.domain.entities.Accomplishments;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
-public interface AccomplishmentsRepository extends JpaRepository<Accomplishments, Integer> {
+public interface AccomplishmentsRepository extends JpaRepository<Accomplishments, Long> {
 }
