+++ src/main/java/com/diplomado/springboot/repositories/EquipmentRepository.java
@@ -3,5 +3,5 @@ package com.diplomado.springboot.repositories;
 import com.diplomado.springboot.domain.entities.Equipment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EquipmentRepository extends JpaRepository<Equipment, Integer> {
+public interface EquipmentRepository extends JpaRepository<Equipment, Long> {
 }
