+++ src/main/java/com/diplomado/springboot/repositories/EvaluationRepository.java
@@ -3,5 +3,5 @@ package com.diplomado.springboot.repositories;
 import com.diplomado.springboot.domain.entities.Evaluation;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EvaluationRepository extends JpaRepository<Evaluation, Integer> {
+public interface EvaluationRepository extends JpaRepository<Evaluation, Long> {
 }
