+++ src/main/java/com/diplomado/springboot/repositories/IncidentsRepository.java
@@ -3,5 +3,5 @@ package com.diplomado.springboot.repositories;
 import com.diplomado.springboot.domain.entities.Incidents;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IncidentsRepository extends JpaRepository<Incidents, Integer> {
+public interface IncidentsRepository extends JpaRepository<Incidents, Long> {
 }
