+++ src/main/java/com/diplomado/springboot/repositories/LaborHistoryRepository.java
@@ -3,5 +3,5 @@ package com.diplomado.springboot.repositories;
 import com.diplomado.springboot.domain.entities.LaborHistory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LaborHistoryRepository extends JpaRepository<LaborHistory, Integer> {
+public interface LaborHistoryRepository extends JpaRepository<LaborHistory, Long> {
 }
