+++ src/main/java/com/diplomado/springboot/repositories/LicensesCertificationsRepository.java
@@ -3,5 +3,5 @@ package com.diplomado.springboot.repositories;
 import com.diplomado.springboot.domain.entities.LicensesCertifications;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LicensesCertificationsRepository extends JpaRepository<LicensesCertifications, Integer> {
+public interface LicensesCertificationsRepository extends JpaRepository<LicensesCertifications, Long> {
 }
