+++ src/main/java/com/diplomado/springboot/repositories/TrainingRepository.java
@@ -3,5 +3,5 @@ package com.diplomado.springboot.repositories;
 import com.diplomado.springboot.domain.entities.Training;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TrainingRepository extends JpaRepository<Training, Integer> {
+public interface TrainingRepository extends JpaRepository<Training, Long> {
 }
