+++ src/main/java/com/diplomado/springboot/repositories/VacationsRepository.java
@@ -3,5 +3,5 @@ package com.diplomado.springboot.repositories;
 import com.diplomado.springboot.domain.entities.Vacations;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VacationsRepository extends JpaRepository<Vacations, Integer> {
+public interface VacationsRepository extends JpaRepository<Vacations, Long> {
 }
